restructure project directories
