+++ src/entities/entity.cpp
@@ -1,6 +1,6 @@
 #include "entities.h"
 #include <string>
-// --------------------------- entity --------------------------- // 
+// --------------------------- entity --------------------------- //
 bool entities::entity::check_collision(const hitbox::hitbox other){
     return body_.get_hitbox().check_collision(other);
 }
@@ -35,13 +35,3 @@ void entities::entity::move_without_event(Vector2 new_position){
     position_ = new_position;
     body_.update_hitboxes(position_);
 }
-
-// --------------------------- builder --------------------------- //
-std::string entities::entity_builder::next_debug_id(const std::string& prefix){
-    auto next_id = debug_id_counts_[prefix];
-    debug_id_counts_[prefix] += 1;
-    return prefix + std::to_string(next_id);
-}
-
-// defining the builder
-entities::entity_builder entities::e_builder;
